nl80211: fix NL80211_ATTR_WDEV type
Bug-Url: https://github.com/svinota/pyroute2/issues/556

+++ pyroute2/netlink/nl80211/__init__.py
@@ -362,7 +362,7 @@ class nl80211cmd(genlmsg):
                ('NL80211_ATTR_INACTIVITY_TIMEOUT', 'hex'),
                ('NL80211_ATTR_RX_SIGNAL_DBM', 'hex'),
                ('NL80211_ATTR_BG_SCAN_PERIOD', 'hex'),
-               ('NL80211_ATTR_WDEV', 'uint32'),
+               ('NL80211_ATTR_WDEV', 'uint64'),
                ('NL80211_ATTR_USER_REG_HINT_TYPE', 'hex'),
                ('NL80211_ATTR_CONN_FAILED_REASON', 'hex'),
                ('NL80211_ATTR_SAE_DATA', 'hex'),
